// src/app/models/client.model.ts
export interface Client {
  ClientCode: string;
  ClientName: string;
  ContactNo: string;
  PANNo: string;
  AadharNo: string;
  DepositBankName: string;
  DepositBankAccountNo: string;
  DepositBankIFSCCode: string;
  WithdrawalBankName: string;
  WithdrawalBankAccountNo: string;
  WithdrawalBankIFSCCode: string;
}